+++ social_flask_mongoengine/models.py
@@ -6,6 +6,7 @@ from social_mongoengine.storage import MongoengineUserMixin, \
                                        MongoengineAssociationMixin, \
                                        MongoengineNonceMixin, \
                                        MongoengineCodeMixin, \
+                                       MongoenginePartialMixin, \
                                        BaseMongoengineStorage
 
 
@@ -14,6 +15,7 @@ class FlaskStorage(BaseMongoengineStorage):
     nonce = None
     association = None
     code = None
+    partial = None
 
 
 def init_social(app, db):
@@ -36,6 +38,11 @@ def init_social(app, db):
         pass
 
     class Code(db.Document, MongoengineCodeMixin):
+        """Mail validation single one time use code"""
+        pass
+
+    class Partial(db.Document, MongoenginePartialMixin):
+        """Partial pipeline storage"""
         pass
 
     # Set the references in the storage class
@@ -43,3 +50,4 @@ def init_social(app, db):
     FlaskStorage.nonce = Nonce
     FlaskStorage.association = Association
     FlaskStorage.code = Code
+    FlaskStorage.partial = Partial