+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Partial pipeline db storage solution
+
 ### Chaged
 - Split from the monolitic [python-social-auth](https://github.com/omab/python-social-auth)
   codebase